bench: official NumSharp-vs-NumPy suite — all-op coverage, 3 sizes, feasible Full-rigor run
Extend the benchmark/ 3-tier suite (BenchmarkDotNet + numpy_benchmark.py + merge) into an
official, reusable, cross-platform NumSharp-vs-NumPy comparison covering ~all supported
np.* ops at three cache-tier sizes (1K / 100K / 10M), with the full per-(op, dtype, N)
ratio matrix.

Pipeline foundation
- BenchmarkConfig.cs: new OfficialBenchmarkConfig — InProcessEmit toolchain + a 50-iteration
  job whose iteration time is capped at 25 ms.
    * InProcessEmit avoids BenchmarkDotNet's out-of-process project search, which fails in
      this repo ("project names need to be unique") because sibling git worktrees under
      .claude/worktrees/ hold same-named copies of the benchmark project. In-process also
      matches the warm long-lived Python/NumPy process, making the cross-language ratio fair.
    * The iteration-time cap is the key feasibility fix: BDN's default Throughput strategy
      ramps to ~8192 invocations/iteration (≈0.5 s/iter) for nanosecond microbenchmarks, so
      a 10M-element op at 50 iterations took ~25 s PER case and the full matrix would take
      days. Capping iteration time lets the pilot pick a per-op invocation count that fits
      25 ms — fast ops still get hundreds–thousands of invocations (tight mean), slow ops
      (argsort@10M) drop to 1/iteration. Measured: a 30-case set went 18 min -> 70 s (~15x)
      with all 50 iterations preserved. Restores the console logger ManualConfig drops.
- Program.cs: apply OfficialBenchmarkConfig as the base config; CLI passes only --filter.
- run_benchmark.py (new): cross-platform orchestrator. Builds C#, runs each suite via BDN
  (per-class JSON = resumable), sweeps NumPy across the 3 sizes, merges, archives to
  results/<timestamp>/. The single reusable entry point.
- numpy_benchmark.py: --size all / --cache-sizes sweep all three sizes in one invocation
  (each result tagged with its n); suite execution extracted into run_suites() and looped.
- merge-results.py: keep ALL sizes (was hard-dropping everything but N=10M), key the join by
  (op, dtype, N), add the N column to every table, and emit a C#-only coverage check (ops
  measured with no NumPy match).

New op coverage (C# benchmark classes + name-matched NumPy suites)
- Comparison: == != < > <= >=
- Bitwise: & | ^ invert left_shift right_shift
- Logic: isnan isinf isfinite maximum minimum isclose allclose array_equal; all any
- NaN reductions: nansum nanmean nanmax nanmin nanstd nanvar nanprod nanmedian
  nanpercentile nanquantile
- Statistics: median percentile quantile average ptp count_nonzero
- Sorting/searching: argsort nonzero searchsorted
- Linear algebra: dot, outer, matmul (matmul dim capped so O(M^3) stays bounded)
- Selection: where(cond,x,y), where(cond)
- Unary extras: cbrt reciprocal square negative positive trunc
- Cumulative: cumprod

Bit-rot repair: SimdVsScalarBenchmarks/MinMaxBenchmarks updated to the current Core API
(ILKernelGenerator->DirectILKernelGenerator.Enabled; np.argmin/argmax now return long).

Verified end-to-end on the comparison suite at 3 sizes x 4 dtypes: 72/72 cases matched
(0 no-data), ratios computed (parity at 10M ~0.9-1.1x, ~2-2.8x slower at 100K — the
small/medium alloc/dispatch overhead). Generated reports archived under results/.

+++ benchmark/NumSharp.Benchmark.GraphEngine/Benchmarks/Reduction/MinMaxBenchmarks.cs
@@ -77,9 +77,9 @@ public class MinMaxBenchmarks : TypedBenchmarkBase
 
     [Benchmark(Description = "np.argmin(a)")]
     [BenchmarkCategory("ArgMin")]
-    public int ArgMin() => np.argmin(_a1D);
+    public long ArgMin() => np.argmin(_a1D);
 
     [Benchmark(Description = "np.argmax(a)")]
     [BenchmarkCategory("ArgMax")]
-    public int ArgMax() => np.argmax(_a1D);
+    public long ArgMax() => np.argmax(_a1D);
 }
